+++ test/api/Api-requests.spec.js
@@ -8,17 +8,17 @@ test('verb methods pass to doRequest', () => {
   nonRequestingApi.doRequest = jest.fn()
 
   nonRequestingApi.get('/')
-  expect(nonRequestingApi.doRequest.mock.calls[0]).toEqual(['get', '/', null, null, {}])
+  expect(nonRequestingApi.doRequest.mock.calls[0]).toEqual(['get', '/', null, null])
 
   nonRequestingApi.post('/')
-  expect(nonRequestingApi.doRequest.mock.calls[1]).toEqual(['post', '/', null, null, {}])
+  expect(nonRequestingApi.doRequest.mock.calls[1]).toEqual(['post', '/', null, null])
 
   nonRequestingApi.put('/')
-  expect(nonRequestingApi.doRequest.mock.calls[2]).toEqual(['put', '/', null, null, {}])
+  expect(nonRequestingApi.doRequest.mock.calls[2]).toEqual(['put', '/', null, null])
 
   nonRequestingApi.patch('/')
-  expect(nonRequestingApi.doRequest.mock.calls[3]).toEqual(['patch', '/', null, null, {}])
+  expect(nonRequestingApi.doRequest.mock.calls[3]).toEqual(['patch', '/', null, null])
 
   nonRequestingApi.delete('/')
-  expect(nonRequestingApi.doRequest.mock.calls[4]).toEqual(['delete', '/', null, null, {}])
+  expect(nonRequestingApi.doRequest.mock.calls[4]).toEqual(['delete', '/', null, null])
 })